move weather api client to domain directory
