+++ src/weather_uk/app/app.py
@@ -1,11 +1,13 @@
 from textual.app import App
 
 from weather_uk import config
-from weather_uk.adapters.weather_api import MetOfficeApi
 from weather_uk.app.screens.forecast import ForecastScreen
 from weather_uk.app.screens.locations import LocationsScreen
 from weather_uk.app.screens.welcome import WelcomeScreen
-from weather_uk.ports.weather_api import AbstractWeatherApi
+from weather_uk.domain.weather_api_client import (
+    AbstractWeatherAPIClient,
+    MetOfficeAPIClient,
+)
 
 
 class WeatherUkApp(App):
@@ -22,7 +24,7 @@ class WeatherUkApp(App):
 
     def on_mount(self) -> None:
         self._user_config: config.UserConfig = config.load_config()
-        self._weather_api: AbstractWeatherApi = MetOfficeApi()
+        self._weather_api: AbstractWeatherAPIClient = MetOfficeAPIClient()
         self._location_id: int | None = None
 
         if not self._user_config.api_key:
