+++ src/weather_uk/domain/forecast.py
@@ -1,9 +1,9 @@
 from weather_uk.data import models
-from weather_uk.ports.weather_api import AbstractWeatherApi
+from weather_uk.domain.weather_api_client import AbstractWeatherAPIClient
 
 
 def get_forecast(
-    api: AbstractWeatherApi,
+    api_client: AbstractWeatherAPIClient,
     location_id: int,
 ) -> list[models.ForecastDay]:
-    return api.get_forecast(location_id)
+    return api_client.get_forecast(location_id)
