+++ src/weather_uk/domain/locations.py
@@ -1,6 +1,8 @@
 from weather_uk.data import models
-from weather_uk.ports.weather_api import AbstractWeatherApi
+from weather_uk.domain.weather_api_client import AbstractWeatherAPIClient
 
 
-def get_locations_list(api: AbstractWeatherApi) -> list[models.Location]:
-    return api.get_locations_list()
+def get_locations_list(
+    api_client: AbstractWeatherAPIClient,
+) -> list[models.Location]:
+    return api_client.get_locations_list()
